mss文章生成 2/requirements.txt
Pillow>=12,<13
python-docx>=1.2,<2
docxtpl>=0.20,<1
